event and round manager foundations
created an event manager based off of scriptable objects, though maybe using a normal mono behaviour is fine. main thing is decoupling its use and revealing dependencies.
roundEnd is an event (scriptable object) that can be invoked by a script, and passes a dictionary of strings keyed to objects. This would allow for dynamic arguments (if only some listeners should use the message).